requirements.txt
python-dotenv
synapseclient